Fix overflow in year enumeration
Whatever it takes — are fully ready for year 2147483647.

+++ FVNever.Reuse/CopyrightNotice.cs
@@ -238,7 +238,7 @@ public class CopyrightNotice
             {
                 if (current is var (currentStart, currentEnd))
                 {
-                    if (start <= currentEnd + 1)
+                    if (start <= (long)currentEnd + 1)
                     {
                         current = (currentStart, Math.Max(currentEnd, end));
                         continue;
